Step 5, connected Sign Up with activity file
Just added more in description. Didn’t even know I could do this. Also
wanted to say we implemented an AlertDialog if you left any of the
things we were looking for empty. Also I
m not that sure about the trim() function to trim the password. If you
use whitespace in the password then it shouldn’t be trimmed in my
opinion.